small fixes found by static analyzer
- printf formatting fixes
- delete[] fixes
- array out of bounds fixes
- compiler warning fixes
- some asserts added

+++ mclib/stuff/angle.hpp
@@ -298,8 +298,8 @@ namespace Stuff {
 					}
 
 				#else
-					cosine = cos(radian);
-					sine = sin(radian);
+					cosine = cosf(radian);
+					sine = sinf(radian);
 				#endif
 					
 					Check_Object(this);
